refactor(DSL) : create Resume file and move Skills class
