+++ src/test/kotlin/study/Resume.kt
@@ -1,0 +1,5 @@
+package study
+
+fun introduce(initialize: Person.() -> Unit): Person {
+    return Person().apply(initialize)
+}
